Passing through the file tree
1. Two parameters are passed to the main method input.
The first is the path to the directory, the second is
resultFileAbsolutePath is the name of the existing file that will
contain the result.

2. For each file in the path directory and in all its subdirectories, do
the following:

If the file has a length no more than 50 bytes, then for all such files:

2.1. Sort them by file name in ascending order, the path is not taken
into account when sorting.

2.2. Rename resultFileAbsolutePath to 'allFilesContent.txt' (use the
FileUtils.renameFile method, and if necessary, FileUtils.isExist).

2.3. In allFilesContent.txt, consistently write down the contents of all
the files from 2.2.1. After each body of the file write "\ n".

All files have the extension txt.